Extract test support to new tRPC router
