+++ apps/nextjs/src/middlewares/auth.middleware.ts
@@ -50,7 +50,7 @@ if (
   process.env.NODE_ENV === "development" ||
   process.env.VERCEL_ENV === "preview"
 ) {
-  publicRoutes.push("/api/trpc/main/testSupport.prepareUser");
+  publicRoutes.push("/api/trpc/test-support/(.*)");
 }
 
 const isPublicRoute = createRouteMatcher(publicRoutes);
