+++ packages/api/index.ts
@@ -1,5 +1,7 @@
 export { oakAppRouter } from "./src/router";
 export type { AppRouter } from "./src/router";
 
+export { testSupportRouter } from "./src/router/testSupport";
+
 export { createContext, createContextWithCustomAuth } from "./src/context";
 export type { Context } from "./src/context";
