+++ packages/api/src/router/testSupport/index.ts
@@ -1,0 +1,8 @@
+import { router } from "../../trpc";
+import { prepareUser } from "./prepareUser";
+
+export const testSupportRouter = router({
+  prepareUser,
+});
+
+export type TestSupportRouter = typeof testSupportRouter;
